fix(web): fix tailwindcss and scripts issues

+++ apps/web/src/domains/courses/modals/CourseDeleteModal.vue
@@ -10,9 +10,12 @@ const props = defineProps<{ isOpen: boolean; courseId: string }>();
 const { loading, execute } = useApi();
 
 const handleSubmit = async () => {
+  if (loading.value) return;
+
   try {
     await execute(() => coursesService.delete(props.courseId), {
       delay_ms: 500,
+      successMessage: "Formation supprimée !",
     });
 
     emit("close");
@@ -42,16 +45,16 @@ const handleCloseModal = () => {
     <div
       v-if="props.isOpen"
       class="fixed inset-0 z-50 flex items-center justify-center bg-black/60 p-4 backdrop-blur-sm"
-      @click.self="emit('close')"
+      @click.self="handleCloseModal()"
     >
       <Transition
         appear
-        enter-active-class="transition duration-300 ease-out delay-75"
-        enter-from-class="opacity-0 scale-95 translate-y-4"
-        enter-to-class="opacity-100 scale-100 translate-y-0"
+        enter-active-class="transition duration-300 ease-out"
+        enter-from-class="opacity-0 scale-95"
+        enter-to-class="opacity-100 scale-100"
         leave-active-class="transition duration-200 ease-in"
-        leave-from-class="opacity-100 scale-100 translate-y-0"
-        leave-to-class="opacity-0 scale-95 translate-y-4"
+        leave-from-class="opacity-100 scale-100"
+        leave-to-class="opacity-0 scale-95"
       >
         <div
           class="w-full max-w-lg overflow-hidden rounded-2xl border border-zinc-700 bg-zinc-800 shadow-2xl shadow-black/50"
